Tuning Type and Option to System Default Group

+++ resources/views/pages/tuning-types/groups/index.blade.php
@@ -45,9 +45,11 @@
                   <td @if($entry->is_default) style="font-weight: bold" @endif>{{ $entry->name }}</td>
                   <td>{{ $entry->tuningTypes()->count() }} Types</td>
                   <td class="td-actions">
-                    <a class="btn btn-icon btn-primary" href="{{ route($route_prefix.'tuning-types.group.edit', ['id' => $entry->id]) }}" title="Edit">
-                      <i data-feather="edit"></i>
-                    </a>
+                    {{-- @if (!$entry->is_system_default) --}}
+                      <a class="btn btn-icon btn-primary" href="{{ route($route_prefix.'tuning-types.group.edit', ['id' => $entry->id]) }}" title="Edit">
+                        <i data-feather="edit"></i>
+                      </a>
+                    {{-- @endif --}}
                     @if ($entry->is_default)
                       <a
                         class="btn btn-icon btn-dark"
@@ -61,7 +63,9 @@
                         <i data-feather="check-circle"></i>
                       </a>
                     @endif
-                    <a class="btn btn-icon btn-danger" onclick="onDelete(this)" data-id="{{ $entry->id }}" title="Delete"><i data-feather="trash-2"></i></a>
+                    @if (!$entry->is_system_default)
+                      <a class="btn btn-icon btn-danger" onclick="onDelete(this)" data-id="{{ $entry->id }}" title="Delete"><i data-feather="trash-2"></i></a>
+                    @endif
                   <form action="{{ route($route_prefix.'tuning-types.group.destroy', $entry->id) }}" class="delete-form" method="POST" style="display:none">
                     <input type="hidden" name="_method" value="DELETE">
                     <input type="hidden" name="_token" value="{{ csrf_token() }}">
